refactor(ui): extract VersionEntry and groupVersionHistory from InventoryVersions
